requirements.txt
fastapi
uvicorn
openai
networkx
pyyaml
python-dotenv